requirements .txt
streamlit==1.37.1
pymupdf==1.24.10
sentence-transformers==3.1.1
faiss-cpu==1.8.0
ibm-watsonx-ai==1.1.15
hugchat==0.3.1
transformers==4.44.2
pillow==10.4.0
pandas==2.2.3
langdetect==1.0.9
pytesseract==0.3.13
python-dotenv==1.0.1